dried logic pulling product from sidebar
